requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1
